# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Adaptive Math API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostic/start:
    post:
      operationId: startDiagnostic
      tags: [diagnostic]
      summary: Start a diagnostic assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDiagnosticBody"
      responses:
        "200":
          description: Diagnostic session with first question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostic/{sessionId}/answer:
    post:
      operationId: submitDiagnosticAnswer
      tags: [diagnostic]
      summary: Submit answer during diagnostic
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: Updated session with next question or completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostic/{sessionId}/results:
    get:
      operationId: getDiagnosticResults
      tags: [diagnostic]
      summary: Get diagnostic assessment results and ability map
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diagnostic results with cognitive report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticResults"
...
  /practice/next-question:
    get:
      operationId: getNextQuestion
      tags: [practice]
      summary: Get next adaptive question for current topic
      parameters:
        - name: topicId
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
            enum: [DIAGNOSTIC, CALIBRATION, MASTERY, REVIEW]
      responses:
        "200":
          description: Next question with metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeQuestion"
  /practice/answer:
    post:
      operationId: submitPracticeAnswer
      tags: [practice]
      summary: Submit answer for a practice question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAnswerBody"
      responses:
        "200":
          description: Answer result with misconception detection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAnswerResult"
  /practice/session-stats:
    get:
      operationId: getPracticeSessionStats
      tags: [practice]
      summary: Get current practice session statistics
      parameters:
        - name: topicId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session stats including streak, accuracy, state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeSessionStats"
...
  /review/answer:
    post:
      operationId: submitReviewAnswer
      tags: [review]
      summary: Submit answer for a review item (updates SM-2 schedule)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAnswerBody"
      responses:
        "200":
          description: Updated review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"
...
components:
  schemas:
...
    StudentAbility:
      type: object
      properties:
        topicId:
          type: string
        topicName:
          type: string
        theta:
          type: number
          description: IRT ability score (-3 to +3)
        masteryPercent:
          type: integer
        state:
          type: string
          enum: [NOT_STARTED, DIAGNOSTIC, CALIBRATION, MASTERY, REVIEW]
        lastPracticedAt:
          type: string
          nullable: true
      required:
        - topicId
        - topicName
        - theta
        - masteryPercent
        - state
    StudentMisconception:
      type: object
      properties:
        id:
          type: string
        topicId:
          type: string
        topicName:
          type: string
        errorTag:
          type: string
        description:
          type: string
        occurrences:
          type: integer
        isResolved:
          type: boolean
        detectedAt:
          type: string
      required:
        - id
        - topicId
        - topicName
        - errorTag
        - description
        - occurrences
        - isResolved
        - detectedAt
...
    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        latex:
          type: string
          nullable: true
          description: LaTeX representation of the question
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        topicId:
          type: string
        difficulty:
          type: number
        difficultyLabel:
          type: string
          enum: [EASY, NORMAL, HARD, VERY_HARD]
      required:
        - id
        - text
        - options
        - topicId
        - difficulty
        - difficultyLabel
    QuestionOption:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        latex:
          type: string
          nullable: true
        isCorrect:
          type: boolean
        misconceptionTag:
          type: string
          nullable: true
          description: Error tag if this option represents a common misconception
      required:
        - id
        - text
        - isCorrect
    StartDiagnosticBody:
      type: object
      properties:
        grade:
          type: integer
      required:
        - grade
    DiagnosticSession:
      type: object
      properties:
        sessionId:
          type: string
        currentQuestion:
          $ref: "#/components/schemas/Question"
          nullable: true
        questionIndex:
          type: integer
        totalQuestions:
          type: integer
        isComplete:
          type: boolean
        topicsAssessed:
          type: array
          items:
            type: string
      required:
        - sessionId
        - questionIndex
        - totalQuestions
        - isComplete
        - topicsAssessed
    SubmitAnswerBody:
      type: object
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
        responseTimeMs:
          type: integer
      required:
        - questionId
        - selectedOptionId
        - responseTimeMs
    DiagnosticResults:
      type: object
      properties:
        sessionId:
          type: string
        abilities:
          type: array
          items:
            $ref: "#/components/schemas/StudentAbility"
        weakTopics:
          type: array
          items:
            type: string
        strongTopics:
          type: array
          items:
            type: string
        cognitiveReport:
          $ref: "#/components/schemas/CognitiveReport"
      required:
        - sessionId
        - abilities
        - weakTopics
        - strongTopics
        - cognitiveReport
...
    PracticeQuestion:
      type: object
      properties:
        question:
          $ref: "#/components/schemas/Question"
        learningState:
          type: string
          enum: [DIAGNOSTIC, CALIBRATION, MASTERY, REVIEW]
        targetProbability:
          type: number
          description: Target success probability (0-1), aim for 0.75
        sessionProgress:
          type: integer
          description: Questions answered in this session
        hintAvailable:
          type: boolean
      required:
        - question
        - learningState
        - targetProbability
        - sessionProgress
        - hintAvailable
    PracticeAnswerBody:
      type: object
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
        topicId:
          type: string
        responseTimeMs:
          type: integer
      required:
        - questionId
        - selectedOptionId
        - topicId
        - responseTimeMs
    PracticeAnswerResult:
      type: object
      properties:
        isCorrect:
          type: boolean
        correctOptionId:
          type: string
        explanation:
          type: string
        misconception:
          $ref: "#/components/schemas/StudentMisconception"
          nullable: true
        updatedTheta:
          type: number
        supplementaryRound:
          type: boolean
          description: True if misconception detected and supplementary questions will be inserted
        newLearningState:
          type: string
          enum: [DIAGNOSTIC, CALIBRATION, MASTERY, REVIEW]
          nullable: true
        masteryAchieved:
          type: boolean
      required:
        - isCorrect
        - correctOptionId
        - explanation
        - updatedTheta
        - supplementaryRound
        - masteryAchieved
    PracticeSessionStats:
      type: object
      properties:
        topicId:
          type: string
        learningState:
          type: string
          enum: [DIAGNOSTIC, CALIBRATION, MASTERY, REVIEW]
        questionsAnswered:
          type: integer
        correctCount:
          type: integer
        accuracyPercent:
          type: integer
        currentTheta:
          type: number
        recentStreak:
          type: integer
        veryHardCorrect:
          type: integer
        openMisconceptions:
          type: integer
        timeSpentMinutes:
          type: number
      required:
        - topicId
        - learningState
        - questionsAnswered
        - correctCount
        - accuracyPercent
        - currentTheta
        - recentStreak
        - veryHardCorrect
        - openMisconceptions
        - timeSpentMinutes
...
    ReviewItem:
      type: object
      properties:
        id:
          type: string
        topicId:
          type: string
        topicName:
          type: string
        dueDate:
          type: string
        interval:
          type: integer
          description: Current SM-2 interval in days
        repetitions:
          type: integer
        easeFactor:
          type: number
          description: SM-2 ease factor (>=1.3)
        status:
          type: string
          enum: [DUE, UPCOMING, COMPLETED]
      required:
        - id
        - topicId
        - topicName
        - dueDate
        - interval
        - repetitions
        - easeFactor
        - status
    ReviewAnswerBody:
      type: object
      properties:
        reviewItemId:
          type: string
        questionId:
          type: string
        selectedOptionId:
          type: string
        responseTimeMs:
          type: integer
        qualityScore:
          type: integer
          minimum: 0
          maximum: 5
          description: SM-2 quality score (0-5)
      required:
        - reviewItemId
        - questionId
        - selectedOptionId
        - responseTimeMs
        - qualityScore
...
    CognitiveReport:
      type: object
      properties:
        responseTimeAvgMs:
          type: integer
        confidenceLevel:
          type: number
          description: 0-100 score
        patternRecognition:
          type: number
          description: 0-100 score
        conceptualUnderstanding:
          type: number
          description: 0-100 score
        adaptabilityScore:
          type: number
          description: 0-100 score
        persistenceIndex:
          type: number
          description: 0-100 score
        totalQuestionsAnswered:
          type: integer
        accuracyOverall:
          type: number
          description: 0-100 percent
        topMisconceptions:
          type: array
          items:
            $ref: "#/components/schemas/StudentMisconception"
      required:
        - responseTimeAvgMs
        - confidenceLevel
        - patternRecognition
        - conceptualUnderstanding
        - adaptabilityScore
        - persistenceIndex
        - totalQuestionsAnswered
        - accuracyOverall
        - topMisconceptions