# lib/api-zod/src/generated/types/practiceAnswerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adaptive Math API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeAnswerBody {
  questionId: string;
  selectedOptionId: string;
  topicId: string;
  responseTimeMs: number;
}