# lib/api-zod/src/generated/types/practiceQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adaptive Math API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeQuestionLearningState } from "./practiceQuestionLearningState";
import type { Question } from "./question";

export interface PracticeQuestion {
  question: Question;
  learningState: PracticeQuestionLearningState;
  /** Target success probability (0-1), aim for 0.75 */
  targetProbability: number;
  /** Questions answered in this session */
  sessionProgress: number;
  hintAvailable: boolean;
}