# lib/api-zod/src/generated/types/submitAnswerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Adaptive Math API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitAnswerBody {
  questionId: string;
  selectedOptionId: string;
  responseTimeMs: number;
}